initial implementation of a CSS selector matching optimization
The idea is to avoid repeated checks of CssSimpleSelector against the
same part of the doctree.
E.g

.navigation * { background-color:green }

Would result in checking for class="navigation" all the way down to the
document root for all elements.
The optimization shortcuts this, for parts of the doctree that have been
checked before.