Implement some keystrokes as shortcuts to option menuitems
Jump to the next menuitem whose label starts with the pressed key.
If there's more than one item starting with the same letter, go to
the next one, then wrap.

This works when the OPTION menu is focused, not displayed.
FLTK has a hacked way of handling menus which will change in
FLTK3 [1] [2].

Making it work when the menu is displayed would involve copying a lot of
code and dealing with hackish FLTK internals, which is not stable. It
seems wiser to use this workaround until FLTK3 solves the problem.

[1] http://fltk.org/newsgroups.php?s1060+gfltk.general+v1070+T0+Qmenu+scrolling
[2] http://fltk.org/newsgroups.php?s1080+gfltk.general+v1087+T0+Qmenu+scrolling