html5 character references
some of them are more than one unicode codepoint. I know there were cases
that resulted in six bytes, which could in principle be longer than the
reference that it's replacing (e.g., "&amp;" is five bytes). So this
required some reworking.

As for (mostly) duplicating the strings for html4 and html5, &lang; and
&rang; are currently different and it would have felt a little silly in
2014 to have, say, a bool for whether an html5 ref appears in html4 plus
special code for lang/rang.