fix the icon name fix
Jorge found that AddressSanitizer complained about heap-use-after-free.

The problem was that, with window name and icon name set to the same
copied string (there is no separate copy_label() for the icon name),
when Fl_Window::copy_label() was called to set the copied label to a
new value, the old string is free()d and then iconlabel() is used
before we have a chance to set it to our newly copied string.