Fix for updateReference() (missing content, overlapped text sometimes)
This is a race condition. When loading [1], tables under the form date camp,
may dissapear and some overlapping text be shown below.

How to reproduce: load the site (network effect may be enough), or
go back and forward a few times, or maximize then back and forward.

[1] http://www.btgpactual.mdgms.com/www/chile/resume.html

FWIW this is the only regression I've found so far.