Add scroll_step option
When using the mouse wheel to scroll a page, the default scroll step was
causing a very slow scrolling speed. The new option "scroll_step" allows
the user to define how many pixels the page is moved in every step of
the mouse wheel. The default is increased to 100 pixels per step.