Add detection logic for OpenSSL and mbedTLS
By default we first look for OpenSSL and if not found we try to find
mbedTLS.

It is an error to build without a suitable TLS library, unless the
search is skipped with --disable-tls. In that case, the support for TLS
is not built into Dillo.

This prevents accidental errors at configure time that end up disabling
TLS without the user knowledge.