Search for libiconv before libc
In BSD systems where libiconv provides /usr/local/include/iconv.h and
libc provides /usr/include/iconv.h there is a problem when we can
use the headers of libiconv but end up linking with libc. The workaround
is to search first for libiconv, and if not found assume it is safe to
link with the libc.