Constraint width in calcFinalWitdh()
Then min-width or max-width is given in the CSS style for the current
widget, the returned width should be clamped by those values.

The condition *finalValue != 1 was used to clamp with min-width, but it
doesn't capture the situation when the finalValue was not yet set.

Similarly, for the max-width, the finalWidth was only set when
*finalValue == -1, which prevented the clamping when the width is set.

The change always perform the clamp when finalWidth is not set or when
the value exceeds the min or max value. When both min-width and
max-width are given but width is -1, the resulting width is the
min-witdh, as the max-width clamp rule won't trigger after setting the
*finalValue to the min-witdh.