Ignore percent width td attributes in tables
The width attribute in tr and td tags is deprecated in HTML 4.01 and
obsolete in HTML 5. However, we should still support it when loading the
page in the transitional mode.

Currently, the relative values of the width attribute are transformed
into CSS values. They cause the available text width to be computed by
applying the relative value to the cell width, which causes an
unintended effect.

The workaround for now is to simply ignore the value when it is
specified as a percent.