Set forceValue when correcting a requisition
For cases where the parent widget has the width set to auto we need to
go to the above parent to compute the available width. This is required
to make tables with a width of 100% occupy the whole available width of
the parent.

In particular, fixes render/table-max-width.html test, which was relying
on a {-1,-1,-1} requisition to have a value different than -1 in the
width to consider that the CSS width is effective.