Handle .exe extension in Windows for dpis
On Cygwin the binaries have the ".exe" extension, so the logic to detect
dpi files has to take it into account when scanning the dpi_dir. The
extension DPI_EXT is defined to ".dpi" or ".dpi.exe" depending on the
platform extension discovered by autoconf.

Similarly, the dpidrc configuration file is also configured to match the
name of the DPI binaries.

See: https://github.com/cygwinports-extras/dillo/blob/master/3.0.2-exeext.patch
Fixes: https://github.com/dillo-browser/dillo/issues/105