Improve quirk for text/xhtml content type
When a <meta> tag reports the "text/xhtml" content, we were correcting
it to the type guessed in TypeDet. However, the current
implementation to guess XHTML and HTML pages fails if the doctype is not
at the start of the document, falling back to text/plain.

A more robust solution is to set the TypeNorm to
"application/xhtml+xml", which can be handled by a_Mime_get_viewer() as
an HTML-like document.

Reported-by: Kevin Koster <dillo@ombertech.com>
See: https://lists.mailman3.com/hyperkitty/list/dillo-dev@mailman3.com/thread/7GJ4AAMFFPEHOIYEOH4NHVMSXMJDFYXG/