Define CssLength as struct instead of int
The int type doesn't have a fixed size, and is only guarantee to hold 16
bits. The current implementation assumes a size of at least 32 bits, an
uses three bits to encode the type of information stored in the rest.

To add more types of lengths we would need to take more bits from the
value itself.

A simpler approach is just to use a enumeration to take care of the type
of length and a union to encapsulate the different lengths values.