Add support for more CSS units
Implements support for ch, rem, vw, vh, vmin and vmax units of CSS
lengths. For now the units relative to the viewport are only computed
once, and they won't change when the window is resized, but only when
the page is reloaded.

See: https://www.toomanyatoms.com/software/mobilized_dillo.html
Authored-By: dogma