Generate version including git commit
Use git describe to get the current commit and other information to set
VERSION in config.h, when we are not using a tarball release. The file
".tarball-version" stores the version inside the tarball, so it is fixed
for releases. The .version file is not used, so no need to generate it.

If the git program is not present when running autogen.sh, the version
will be set to UNKNOWN. This situation will be very rare as you will
need git to fetch the sources, but continues to allow building Dillo
without the git program.

Fixes: https://github.com/dillo-browser/dillo/issues/288