Add git commit to the reported version with -v
When git is available, records the git commit in the dillo binary, so
dillo -v reports the version and the commit. Otherwise only the release
version is reported (as we do now).

The commit is *always* recomputed when "make" is called. This allows
switching branches, doing a dirty rebuild and still get the correct
commit with "dillo -v".

Using AC_INIT() doesn't update the commit when the source is already
configured.

Fixes: https://github.com/dillo-browser/dillo/issues/288