Print error if dpi_dir cannot be opened
On cases where the system directory cannot be opened, dpid will not load
the builtin plugins causing Dillo to fail to work properly with file: or
data: URLs. Reporting a failure to open the directory allows users to
determine the cause of the problem.

See: https://github.com/dillo-browser/dillo/issues/337