Add about:cache page with cache statistics
The internal network cache holds entries for all the URLs fetched by
Dillo, and it only grows over time. It is convenient to be able to see
what elements are being currently stored so we can have choose a better
criteria if we decide to remove entries to avoid a huge memory usage
over time.

The entries are generated every time the page is requested, so it always
reflects the current state.