Detect leaks in test suite with Asan
Leaks from libfontconfig.so are excluded for now, as it doesn't seem to
be our fault. The gcc mechanism to suppress leaks doesn't seem to be
working, so we fallback to awk.

In the presence of leaks, that specific test will fail by returning
non-zero, but this will not be what we want for those tests that are
currently failing with XPASS. We will only consider the ones that are
passing as those targeted for leak detection.