Avoid new Embed or Resource on invisible inputs
When an input element should not be displayed due to display:none, don't
allocate the Resource or the Embed directly. It won't be added to the
Widget tree either. This prevents allocating Resources that are not tied
to a Widget so they are not leaked when the tree is destroyed.

The Embed or Resource elements won't be needed as the DilloHtmlInput
handles all the form logic.