Fix OAuth by allowing cookies in root redirects
For OAuth to work the authentication server sets a cookie and then
performs a 30X redirect to the site trying to authenticate. This
cookie was being rejected due to the organization being different.

To continue blocking third-party cookies for tracking image pixels, we
use the requester and WEB_RootUrl flag to discern if the request is
caused by a root redirect or by a resource. Only root redirect cookies
are allowed to be set.

Fixes: https://bug.dillo-browser.org/253/
Reviewed-by: dogma