Connect to the first working control socket
When dillo crashes or is killed the control socket won't be erased,
causing multiple sockets to remain in the control directory. Instead of
requesting the user to specify the PID manually when there are multiple
entries, remove those that don't work because dillo is no longer
listening and only require the PID when there are multiple active
control sockets.