Create and own hidden resources in DilloHtmlInput
Instead of avoiding the creation of resources for elements with
display:none (or descendants), always create them but set the owner to
the DilloHtmlInput so they are destroyed alongside the form inputs.

Prevents the leakage of the resources while we can continue to use all
the access methods in DilloHtmlForm::buildQueryData() for each type of
resource. Likely introduced by cfa3923a5ffed809b4d4a54f6100f85b9f7a296d.

Notice that the two options to create hidden elements are not the same:

  <input type=hidden>
  <input style="display:none">

The former defines a text input that will be revealed when the user
decides to show hidden inputs. The latter can be of different input
types and will only be shown if the CSS rules are ignored. In both cases
the content of the input must be sent with the form, regardless of their
visibility.

Fixes: https://bug.dillo-browser.org/506/