Wait until the page has fully rendered
When waiting with `dilloc wait` make sure we only send the ready signal
after the page has fully rendered. We need to run the last layout pass
and flush the output to the screen.

For now we directly call the generalIdle and Fl::flush so that we can
guarantee the canvas is updated. In the future, we should switch to an
emitted event so we don't need to reach into FltkPlatform.