Implement non-blocking control socket
When using dilloc in a pipe to perform a filter operation like:

  $ dilloc dump | filter | dilloc load

There is a chance that the dump command begins the execution before all
the data is dumped from dilloc dump. This causes dillo to hang as soon
as it starts handling the load command, which is blocking, and no
progress can be done in the dump side.

To prevent deadlocks both in the read or write side, all operations are
now non-blocking. The UI will continue to be responsive, even if dillo
is loading the current page.

Fixes: https://bug.dillo-browser.org/515/