Add border in images without a buffer
When loading a page with images disabled that mixes images and text, it
is hard to determine which parts belong to the alt-text of the image and
which is the surrounding text. To make it easier to differentiate, add a
border around the alt-text image until the real image is loaded.